backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
sqlalchemy
psycopg2-binary
pandas
openpyxl
lxml
beautifulsoup4
